backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
social-auth-app-django==5.4.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
Jinja2==3.1.2
django-jinja==2.11.0

